Fix some bugs, and Application window resize callback

+++ Pika/src/Pika/Renderer/Camera2DController.cpp
@@ -49,8 +49,8 @@ namespace Pika {
 
 	bool Camera2DController::onWindowResizeEvent(WindowResizeEvent& vEvent)
 	{
-		float Width = vEvent.getWidth();
-		float Height = vEvent.getHeight();
+		float Width = static_cast<float>(vEvent.getWidth());
+		float Height = static_cast<float>(vEvent.getHeight());
 		m_AspectRatio = Width / Height;
 		m_Camera.setProjectionMatrix(-m_AspectRatio * m_ZoomLevel, m_AspectRatio * m_ZoomLevel, -m_ZoomLevel, m_ZoomLevel);
 		return false;
